Gemini-GPT/requirements.txt
streamlit==1.37.1
google-generativeai
